#10801 Refactor: Extracted new HistoryTablesTest
